# agent/api/v1/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for managing btrfs volumes, snapshots, clones, NFS exports, and background tasks.",
        "title": "btrfs-nfs-csi Agent API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/v1/snapshots": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all snapshots. Use detail=true for full metadata. Supports pagination and label filtering.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "snapshots"
                ],
                "summary": "List snapshots",
                "parameters": [
                    {
                        "enum": [
                            "true"
                        ],
                        "type": "string",
                        "description": "Return full detail",
                        "name": "detail",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Pagination cursor",
                        "name": "after",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "Label filter (key=value), repeatable",
                        "name": "label",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SnapshotListResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a read-only btrfs snapshot of a volume. Returns existing snapshot on conflict.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "snapshots"
                ],
                "summary": "Create a snapshot",
                "parameters": [
                    {
                        "description": "Snapshot configuration",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SnapshotCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.SnapshotDetailResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Snapshot already exists",
                        "schema": {
                            "$ref": "#/definitions/models.SnapshotDetailResponse"
                        }
                    }
                }
            }
        },
        "/v1/snapshots/{name}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns detailed metadata for a single snapshot.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "snapshots"
                ],
                "summary": "Get a snapshot",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Snapshot name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SnapshotDetailResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes a snapshot and its btrfs subvolume.",
                "tags": [
                    "snapshots"
                ],
                "summary": "Delete a snapshot",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Snapshot name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/v1/volumes/{name}/snapshots": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns snapshots for a specific volume. Supports detail, pagination, and label filtering.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "snapshots"
                ],
                "summary": "List snapshots for a volume",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Volume name",
                        "name": "name",
                        "in": "path"
                    },
                    {
                        "enum": [
                            "true"
                        ],
                        "type": "string",
                        "description": "Return full detail",
                        "name": "detail",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Pagination cursor",
                        "name": "after",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "Label filter (key=value), repeatable",
                        "name": "label",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SnapshotListResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                }
            }
        },
...
        "models.SnapshotCreateRequest": {
            "type": "object",
            "properties": {
                "labels": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "name": {
                    "description": "^[a-zA-Z0-9_-]{1,128}$",
                    "type": "string"
                },
                "volume": {
                    "description": "source volume name",
                    "type": "string"
                }
            }
        },
        "models.SnapshotDetailResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "exclusive_bytes": {
                    "type": "integer"
                },
                "labels": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "readonly": {
                    "type": "boolean"
                },
                "size_bytes": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "used_bytes": {
                    "type": "integer"
                },
                "volume": {
                    "type": "string"
                }
            }
        },
        "models.SnapshotListResponse": {
            "type": "object",
            "properties": {
                "next": {
                    "type": "string"
                },
                "snapshots": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.SnapshotResponse"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "models.SnapshotResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "size_bytes": {
                    "type": "integer"
                },
                "used_bytes": {
                    "type": "integer"
                },
                "volume": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Tenant token as \"Bearer \u003ctoken\u003e\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}